backend/requirements.txt
docxtpl==0.6.3
python_gitlab==1.8.0
requests==2.22.0
Django==1.11.8
Jinja2==2.10.1
toml==0.10.0
python_docx==0.8.7
PyMySQL==0.9.3
docx==0.2.4
djangorestframework==3.10.3
